Detect Spatial and Environmental Outliers in Occurrence Data

This function identifies spatial and environmental outliers in species occurrence data. It can apply different statistical methods to detect geographic outliers and environmental anomalies.

Usage

findOutliers(
  pres,
  spatial = TRUE,
  environmental = TRUE,
  method = "grubbs",
  distEnvMethod = "euclidean",
  scaleData = TRUE,
  pval = 1e-05,
  checkPairs = FALSE,
  kRosner = NULL,
  verbose = TRUE
)

Arguments

pres

An `sf` object with POINT geometry representing species occurrence data.

spatial

Logical; if `TRUE`, performs spatial outlier detection.

environmental

Logical; if `TRUE`, performs environmental outlier detection.

method

Character; method for detecting outliers. Options are 'grubbs', 'iqr', 'dixon', or 'rosner'. Only a single method can be used at a time.

distEnvMethod

Character; method for calculating environmental distances. Options are 'euclidian', 'manhattan', 'cosine, or 'mahalanobis'.

scaleData

Logical; if `TRUE`, scales environmental data before calculating distances.

pval

Numeric; user-specified p-value for assessing the significance of the test statistic.

checkPairs

Logical; if `TRUE`, checks for a pair of outliers using the Grubbs test (only applicable for sample sizes < 30).

kRosner

Integer; number of outliers to test for with the Rosner test. Applicable only when `method = 'rosner'`.

verbose

Logical; if `TRUE`, prints messages about the number of d detected outliers.

Details

The function takes as input an `sf` object with POINT geometry and applies spatial and/or environmental outlier detection methods. Users can choose between multiple outlier detection methods, including Grubbs, Dixon, IQR, or Rosner. For environmental outlier detection, the function calculates the distance between occurrences based on either Euclidian, Manhattan, Cosine, or Mahalanobis distance.

Author

Cory Merow <cory.merow@gmail.com>, Gonzalo E. Pinilla-Buitrago